Refactored taosupport
