+++ go/apps/simpleexample/SimpleDomain/copybins
@@ -1,9 +1,9 @@
 #
-
 cp ~/bin/SimpleDomainService /Domains
 cp ~/bin/SimpleClient /Domains
 cp ~/bin/linux_host /Domains
 cp ~/bin/QuoteServer /Domains
 cp ~/bin/SimpleServer /Domains
 cp ~/bin/domain_server /Domains
-cp ~/bin/simpleclient_cc.exe /Domains
+# This should already be there
+# cp ~/bin/simpleclient_cc.exe /Domains
